Implemented the logic controlling lowering the stack after lines clear animationImplemented the logic controlling lowering the stack after lines clear animationImplemented the clearing the top rows after lowering the stack
Each time the stack is lowered after clearing the line, the first visible row is unmodified, so it must be cleared. Additionally, in order not to leave garbage in invisible rows (which could prevent spawning and movement of the next piece, which is a known bug in the NES Tetris), invisible lines are also cleaned.