Method to freezing the game thread between frames restored
"SDL_Sleep(1)" cannot be used, as it consumes all the power of the CPU core. Apparently this feature uses shitty busy waiting, so a better solution is to calculate the number of milliseconds to wait and call "SDL_Delay" once (and use the rest of the time on busy waiting).